Initial implementation of: reader thread for async handling of rig data, UDP multicast publisher routine and rig state poll thread routine. The reader thread can correctly handle asynchronous data, such as transceive or spectrum data. Work in progress: multi-platform code for I/O routines still missing and the poll routine is not yet in use. Tested briefly on an IC-7300 so far.

+++ rigs/icom/icom.h
@@ -181,7 +181,7 @@ struct icom_spectrum_scope_cache
     freq_t spectrum_span_freq; /*!< The frequency span of the current spectrum scope line being received */
     freq_t spectrum_low_edge_freq; /*!< The low edge frequency of the current spectrum scope line being received */
     freq_t spectrum_high_edge_freq; /*!< The high edge frequency of the current spectrum scope line being received */
-    int spectrum_data_length;     /*!< Number of bytes of 8-bit spectrum data in the data buffer. The amount of data may vary if the rig has multiple spectrum scopes, depending on the scope. */
+    size_t spectrum_data_length;     /*!< Number of bytes of 8-bit spectrum data in the data buffer. The amount of data may vary if the rig has multiple spectrum scopes, depending on the scope. */
     unsigned char *spectrum_data; /*!< Dynamically allocated buffer for raw spectrum data */
 };
